+++ src/game/game.js
@@ -19,7 +19,7 @@ export class Game {
         this.howToPlayButton = document.getElementById('how-to-play');
         this.howToPlayDialog = document.getElementById('how-to-play-dialog');
         this.DEALER_DELAY_MS = 4e2;
-        this.NEW_ROUND_DELAY_MS = 1e3;
+        this.NEW_ROUND_DELAY_MS = 3e3;
         this.dealer = new Dealer();
         this.player = new Player();
         this.hitEnterKeyHandler = this.hitEnterKeyHandler.bind(this);
@@ -56,9 +56,7 @@ export class Game {
         this.updateButtons(true, false, buttons);
         this.player.addCard();
         if (this.player.score > Rules.BLACK_JACK) {
-            const bust = this.player.bust;
-            this.player.refreshMoneyAfterResult(bust);
-            this.player.setResultMessage(bust);
+            this.player.displayResult(this.player.bust);
             setTimeout(() => this.newRound(), this.NEW_ROUND_DELAY_MS);
             return;
         }
@@ -84,8 +82,7 @@ export class Game {
         else {
             playerResult = this.player.push;
         }
-        this.player.refreshMoneyAfterResult(playerResult);
-        this.player.setResultMessage(playerResult);
+        this.player.displayResult(playerResult);
     }
     stand() {
         this.ripple(this.standButton);